feat: implement OAuth2 data structures for token, user info, and provider configuration
