+++ src/core/oauth/store.rs
@@ -1,0 +1,40 @@
+use serde::{Deserialize, Serialize};
+
+/// Represents an OAuth2 access token and optional refresh token.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct OAuth2Token {
+    pub access_token: String,
+    pub refresh_token: Option<String>,
+    pub expires_in: Option<u64>,
+    pub token_type: Option<String>,
+    pub scope: Option<String>,
+}
+
+/// Represents a generic OAuth2 user info response.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct OAuth2UserInfo {
+    pub provider: OAuth2Provider, // e.g., "google", "github"
+    pub id: String,
+    pub email: Option<String>,
+    pub name: Option<String>,
+    pub avatar_url: Option<String>,
+    pub raw: serde_json::Value,
+}
+
+/// Configuration for an OAuth2 provider.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct OAuth2Config {
+    pub client_id: String,
+    pub client_secret: String,
+    pub auth_url: String,
+    pub token_url: String,
+    pub user_info_url: String,
+    pub redirect_uri: String,
+    pub scopes: Vec<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct OAuth2Provider {
+    pub name: String,
+    pub config: OAuth2Config,
+}
